pyproject.toml
[build-system]
requires = ["setuptools>=68", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "devpack"
version = "0.1.0"
description = "CLI tool that helps developers add agent skills to an existing repo based on the detected tech stack."
readme = "readme.md"
requires-python = ">=3.11"
dependencies = [
    "typer>=0.12",
    "inquirerpy>=0.3",
    "pyyaml>=6.0",
    "claude-agent-sdk>=0.1.38",
]

[project.scripts]
devpack = "devpack.cli:app"

[project.optional-dependencies]
dev = [
    "pytest>=8.0",
    "pytest-cov",
]

[tool.setuptools.packages.find]
where = ["src"]

[tool.pytest.ini_options]
testpaths = ["tests"]
